add @past3lle/carousel code, remove local
