+++ package.json
@@ -34,6 +34,8 @@
     "@metamask/jazzicon": "^2.0.0",
     "@past3lle/assets": "^0.0.10",
     "@past3lle/components": "^0.0.44",
+    "@past3lle/carousel": "^0.0.6",
+    "@past3lle/carousel-hooks": "^0.0.4",
     "@past3lle/constants": "^0.0.17",
     "@past3lle/hooks": "^0.0.12",
     "@past3lle/theme": "^0.0.29",
