+++ src/components/Carousel/ProductCarousels.tsx
@@ -1,13 +1,18 @@
+import {
+  BaseCarouselProps,
+  ButtonCarousel,
+  ButtonCarouselProps,
+  CarouselChildrenProps,
+  HorizontalSwipeCarousel,
+  WithTouchAction,
+} from '@past3lle/carousel'
 import { SmartVideoProps as LazyVideoProps, SmartImageProps, SmartImg } from '@past3lle/components'
-import ButtonCarousel, { ButtonCarouselProps } from 'components/Carousel/common/components/ButtonCarousel'
 import { useCallback } from 'react'
 import { FragmentProductVideoFragment } from 'shopify/graphql/types'
 import { isProductVideo } from 'shopify/utils'
 import { ShopImageSrcSet } from 'types'
 
-import HorizontalSwipeCarousel from './HorizontalSwipeCarousel'
-import { CarouselShowcaseVideo } from './common/common'
-import { BaseCarouselProps, CarouselChildrenProps, WithTouchAction } from './common/types'
+import { CarouselShowcaseVideo } from './CarouselShowcaseVideo'
 
 type CarouselData = ShopImageSrcSet | FragmentProductVideoFragment | undefined
 interface ProductCarousel extends Omit<BaseCarouselProps<CarouselData[]>, 'children'> {
