--- src/components/Carousel/common/components/styleds.ts
@@ -1,129 +1,0 @@
-import { Row } from '@past3lle/components'
-import { BLACK, BLACK_TRANSPARENT, OFF_WHITE } from '@past3lle/theme'
-import { ScrollerContainer } from 'components/ScrollingContentPage/styleds'
-import { Z_INDEXES } from 'constants/config'
-import { transparentize } from 'polished'
-import styled from 'styled-components/macro'
-
-const BaseCarouselStep = styled(Row)<{
-  $width: number
-  $height?: string
-}>`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: ${({ $width }) => $width}px;
-  ${({ $height }) => $height && `height: ${$height};`}
-  overflow: hidden;
-  // TODO: BE SURE THIS ISNT SHITTY
-  align-items: flex-start;
-
-  img {
-    max-width: 100%;
-    // box-shadow: 3px 6px 8px 0px #686868ad;
-  }
-`
-
-export const AnimatedCarouselStep = styled(BaseCarouselStep)``
-export const StaticCarouselStep = styled(BaseCarouselStep)<{
-  $transformAmount: number
-}>`
-  transform: ${({ $transformAmount }) => `translateX(${$transformAmount}px)`};
-
-  // transform one differently than the others
-  transition: ${({ $transformAmount }) =>
-    $transformAmount > 0 ? 'transform 1s ease-in-out;' : 'transform 0.7s ease-out'};
-
-  z-index: ${({ $transformAmount }) => ($transformAmount > 0 ? 1 : 0)};
-`
-
-export const CarouselContainer = styled(ScrollerContainer).attrs({ $isVerticalScroll: false })<{
-  $fixedHeight?: number
-  $fixedWidth?: number
-}>`
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: center;
-  align-items: center;
-  // transform requirement
-  position: relative;
-  overflow: hidden;
-  height: 100%;
-  width: 100%;
-  z-index: 1;
-  ${({ $fixedHeight }) => $fixedHeight && `height: ${$fixedHeight}px;`}
-  ${({ $fixedWidth }) => $fixedWidth && `width: ${$fixedWidth}px;`}
-
-  ${({ theme }) => theme.mediaWidth.upToSmall`
-    > ${StaticCarouselStep} {
-        max-width: 100%;
-        height: auto;
-      }
-  `}
-`
-
-export const CarouselButton = styled.div<{ buttonColor: string }>`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  width: 10%;
-  background-color: ${({ buttonColor }) => transparentize(1, buttonColor)};
-
-  cursor: pointer;
-
-  &:hover {
-    background-color: ${({ buttonColor }) => transparentize(0.7, buttonColor)};
-  }
-
-  transition: background-color 0.2s ease-in-out;
-`
-
-export const CarouselButtonContainer = styled.div`
-  position: absolute;
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-  align-items: center;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-
-  cursor: pointer;
-`
-
-export const CarouselIndicator = styled.div<{ isCurrent: boolean; color?: string }>`
-  background: ${({ isCurrent, theme, color = theme.blackOpaqueMore }) => (isCurrent ? color : 'transparent')};
-  color: ${({ isCurrent }) => (isCurrent ? BLACK : OFF_WHITE)};
-  border: 1px solid ${({ isCurrent, theme, color = theme.blackOpaqueMore }) => (isCurrent ? 'transparent' : color)};
-  border-radius: 1rem;
-  flex: 0 1 12%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1.4rem;
-  font-weight: 400;
-  padding: 0.25rem 0;
-`
-export const CarouselIndicatorWrapper = styled(Row)`
-  opacity: 0.5;
-  position: absolute;
-  bottom: 7%;
-  right: 0;
-  margin: 0;
-  padding: 1rem;
-  gap: 0.5rem;
-  width: 100%;
-  justify-content: flex-end;
-  align-items: stretch;
-  z-index: ${Z_INDEXES.SCROLLER_DIV};
-  // background: ${`linear-gradient(275deg, ${BLACK_TRANSPARENT} 33%, transparent)`};
-  background: transparent;
-
-  &:hover {
-    opacity: 1;
-  }
-
-  transition: opacity, background 0.3 ease-in-out;
-`
