+++ src/components/ScrollingContentPage/index.tsx
@@ -1,10 +1,10 @@
+import { useInfiniteVerticalScroll } from '@past3lle/carousel-hooks'
 import { LoadInViewOptions } from '@past3lle/hooks'
 import { isMobile } from '@past3lle/utils'
 import PastelleIvoryOutlined from 'assets/svg/pastelle-ivory-outlined.svg'
 import { FixedAnimatedLoader } from 'components/Loader'
 import { COLLECTION_MAX_WIDTH, MINIMUM_COLLECTION_ITEM_HEIGHT } from 'constants/config'
 import { STIFF_SPRINGS } from 'constants/springs'
-import { useInfiniteVerticalScroll } from 'hooks/useScrollAnimation'
 import { useCallback, useEffect } from 'react'
 import { Product } from 'shopify/graphql/types'
 import { useIsMobileWindowWidthSize } from 'state/window/hooks'
