--- src/hooks/useScrollAnimation/index.ts
@@ -1,16 +1,0 @@
-import useInfiniteHorizontalScroll from './useInfiniteHorizontalScroll'
-import useInfiniteVerticalScroll from './useInfiniteVerticalScroll'
-import { useLimitedHorizontalSwipe, useLimitedVerticalSwipe } from './useLimitedSwipe'
-import useScrollingAnimationSetup from './utils/useScrollSetup'
-import useScrollZoneRef from './utils/useScrollZoneRef'
-
-export * from './types'
-export * from './utils/utils'
-export {
-  useScrollZoneRef,
-  useLimitedHorizontalSwipe,
-  useLimitedVerticalSwipe,
-  useInfiniteHorizontalScroll,
-  useInfiniteVerticalScroll,
-  useScrollingAnimationSetup,
-}
